Update API tests to first get valid session ID
