--- backend-api/tests/api-tests/utils/getRegisteredClient.ts
@@ -1,27 +1,0 @@
-import {
-  GetSecretValueCommand,
-  SecretsManagerClient,
-} from "@aws-sdk/client-secrets-manager";
-
-export interface ClientDetails {
-  client_id: string;
-  client_secret: string;
-  redirect_uri: string;
-}
-
-async function getRegisteredClients(): Promise<ClientDetails[]> {
-  const secretsManagerClient = new SecretsManagerClient({
-    region: "eu-west-2",
-  });
-  const secretName = `${process.env.TEST_ENVIRONMENT}/clientRegistryApiTest`;
-  const command = new GetSecretValueCommand({
-    SecretId: secretName,
-  });
-  const response = await secretsManagerClient.send(command);
-  return JSON.parse(response.SecretString!);
-}
-
-export async function getFirstRegisteredClient(): Promise<ClientDetails> {
-  const clientsDetails = await getRegisteredClients();
-  return clientsDetails[0];
-}
